+++ src/platform.ts
@@ -2,11 +2,7 @@ import { API, DynamicPlatformPlugin, Logger, PlatformAccessory, Service, Charact
 import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
 import { PLATFORM_NAME, PLUGIN_NAME, DeviceURL } from './settings';
 import { Humidifier } from './Devices/Humidifier';
-import {
-  irdevices,
-  devices,
-  SwitchBotPlatformConfig,
-} from './configTypes';
+import { irdevices, device, SwitchBotPlatformConfig, deviceResponses } from './configTypes';
 
 /**
  * HomebridgePlatform
@@ -23,9 +19,6 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
   public axios: AxiosInstance = axios.create({
     responseType: 'json',
   });
-
-  public sensorData = [];
-  private refreshInterval;
 
   constructor(public readonly log: Logger, public readonly config: SwitchBotPlatformConfig, public readonly api: API) {
     this.log.debug('Finished initializing platform:', this.config.name);
@@ -46,7 +39,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
     // setup axios interceptor to add headers / api key to each request
     this.axios.interceptors.request.use((request: AxiosRequestConfig) => {
-      request.headers.Authorization = this.config.credentials ?.openToken;
+      request.headers.Authorization = this.config.credentials?.openToken;
       request.headers['Content-Type'] = 'application/json; charset=utf8';
       return request;
     });
@@ -91,12 +84,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     this.config.options = this.config.options || {};
 
     if (this.config.options) {
-      this.config.options ?.humidifier;
+      this.config.options?.humidifier;
+    }
+
+    if (this.config.options!.ttl! < 120) {
+      throw new Error('TTL must be above 120 (2 minutes).');
     }
 
     if (!this.config.options.ttl) {
-      this.config.options!.ttl! = 300;
-      this.log.warn('TTL not provided, using default TTL.');
+      this.config.options!.ttl! = 120;
+      this.log.warn('Using Default Refresh Rate.');
     }
 
     if (!this.config.credentials) {
@@ -112,21 +109,37 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
    */
   async discoverDevices() {
     const devices = (await this.axios.get(DeviceURL)).data;
+    if (this.config.devicediscovery) {
+      this.deviceListInfo(devices);
+    } else {
+      this.log.debug(JSON.stringify(devices));
+    }
+    this.log.info(`Total Devices Found: ${devices.body.deviceList.length}`);
     for (const device of devices.body.deviceList) {
-      this.log.info(`Total Devices Found: ${device.length}`);
-      this.log.debug(JSON.stringify(device));
-      if (device.deviceType.startsWith('Humidifier')) {
-        // this.deviceinfo(device);
-        this.log.info('Discovered %s %s - %s', device.deviceType, device.deviceModel, device.userDefinedDeviceName);
-        this.createHumidifier(device, devices);
-      } else if (!device.DeviceModel) {
-        this.log.info('A LLC Device has been discovered with a deviceModel that does not start with T5, D6 or T9');
+      if (this.config.devicediscovery) {
+        this.deviceInfo(device);
+      } else {
+        this.log.debug(JSON.stringify(device));
+      }
+      // For Future Devices
+      switch (device.deviceType) {
+        case 'Humidifier':
+          this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
+          this.createHumidifier(device, devices);
+          break;
+        default:
+          this.log.info(
+            `A SwitchBot Device has been discovered with Device Type: ${device.deviceType}, which is currently not supported.`,
+            'Submit Feature Requests Here: https://git.io/JL14Z,',
+          );
       }
     }
   }
 
-  private async createHumidifier(device, devices) {
-    const uuid = this.api.hap.uuid.generate(`${device.name}-${device.deviceID}-${device.deviceModel}`);
+  private async createHumidifier(device: device, devices: deviceResponses) {
+    const uuid = this.api.hap.uuid.generate(
+      `${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+    );
 
     // see if an accessory with the same uuid has already been registered and restored from
     // the cached devices we stored in the `configureAccessory` method above
@@ -139,23 +152,23 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
         // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
         //existingAccessory.context.firmwareRevision = firmware;
-        await this.api.updatePlatformAccessories([existingAccessory]);
+        this.api.updatePlatformAccessories([existingAccessory]);
         // create the accessory handler for the restored accessory
         // this is imported from `platformAccessory.ts`
-        await new Humidifier(this, existingAccessory, device);
-        this.log.debug(`T9 UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
+        new Humidifier(this, existingAccessory, device);
+        this.log.debug(
+          `Humidifier UDID: ${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+        );
       } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
     } else {
       // the accessory does not yet exist, so we need to create it
-      this.log.info('Adding new accessory:', `${device.name} ${device.deviceModel} ${device.deviceType}`);
-      this.log.debug(
-        `Registering new device: ${device.name} ${device.deviceModel} ${device.deviceType} - ${device.deviceID}`,
-      );
+      this.log.info('Adding new accessory:', `${device.deviceName} ${device.deviceType}`);
+      this.log.debug(`Registering new device: ${device.deviceName} ${device.deviceType} - ${device.deviceId}`);
 
       // create a new accessory
-      const accessory = new this.api.platformAccessory(`${device.name} ${device.deviceType}`, uuid);
+      const accessory = new this.api.platformAccessory(`${device.deviceName} ${device.deviceType}`, uuid);
 
       // store a copy of the device object in the `accessory.context`
       // the `context` property can be used to store any data about the accessory you may need
@@ -165,7 +178,9 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       // create the accessory handler for the newly create accessory
       // this is imported from `platformAccessory.ts`
       new Humidifier(this, accessory, device);
-      this.log.debug(`T9 UDID: ${device.name}-${device.deviceID}-${device.deviceModel}`);
+      this.log.debug(
+        `Humidifier UDID: ${device.deviceName}-${device.deviceId}-${device.deviceType}-${device.hubDeviceId}`,
+      );
 
       // link the accessory to your platform
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
@@ -179,9 +194,11 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     this.log.info('Removing existing accessory from cache:', existingAccessory.displayName);
   }
 
-  public deviceinfo(device: irdevices | devices) {
-    if (this.config.devicediscovery) {
-      this.log.warn(JSON.stringify(device));
-    }
+  public deviceListInfo(devices: deviceResponses) {
+    this.log.warn(JSON.stringify(devices));
+  }
+
+  public deviceInfo(device: irdevices | device) {
+    this.log.warn(JSON.stringify(device));
   }
 }