+++ src/settings.ts
@@ -1,19 +1,19 @@
 /**
  * This is the name of the platform that users will use to register the plugin in the Homebridge config.json
  */
-export const PLATFORM_NAME = 'Switch-Bot';
+export const PLATFORM_NAME = 'SwitchBot';
 
 /**
  * This must match the name of your plugin as defined the package.json
  */
-export const PLUGIN_NAME = 'homebridge-switch-bot';
+export const PLUGIN_NAME = 'homebridge-switchbot-openapi';
 
 /**
- * This is the main url used to access switch-bot API
+ * This is the main url used to access SwitchBot API
  */
 export const AuthURL = 'https://api.switch-bot.com';
 
 /**
- * This is the main url used to access switch-bot API
+ * This is the main url used to access SwitchBot API
  */
 export const DeviceURL = 'https://api.switch-bot.com/v1.0/devices';