+++ CHANGELOG.md
@@ -2,14 +2,16 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/)
 
-## [Beta](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/releases/) (2021-XX-XX)
+## [Version 1.0.0](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/releases/tag/v1.0.0) (2020-12-25)
 
 ### Changes
 
-- Once Open Token is available in the SwitchBot App, we will be able to start beta testing.
+- Offical Release of OpenToken Switchbot API Support.
+- Support for [SwitchBot Humidifiers](https://www.switch-bot.com/products/switchbot-smart-humidifier).
 
-## [Version 0.1.0](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/releases/tag/v0.1.0) (2021-12-19)
+## [Version 0.1.0](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/releases/tag/v0.1.0) (2020-12-19)
 
 ### Changes
 
 - Initial Release.
+- This release will only valid that your Open Token Works.